- Adds link preview to images
- Adds option to manual focus in the camera
- Adds support to switch between front and back cameras during video recording
- Adds basic face filters
- Improves image editor, like emojis or text under a drawing can be moved
- Improves speed after taking a picture  
- Fixes issue with emojis disappearing in the image editor

+++ lib/src/views/memories/memories_photo_slider.view.dart
@@ -9,7 +9,7 @@ import 'package:twonly/src/services/api/mediafiles/upload.service.dart';
 import 'package:twonly/src/utils/log.dart';
 import 'package:twonly/src/utils/misc.dart';
 import 'package:twonly/src/views/camera/camera_preview_components/save_to_gallery.dart';
-import 'package:twonly/src/views/camera/share_image_editor_view.dart';
+import 'package:twonly/src/views/camera/share_image_editor.view.dart';
 import 'package:twonly/src/views/components/alert_dialog.dart';
 import 'package:twonly/src/views/components/media_view_sizing.dart';
 import 'package:twonly/src/views/components/video_player_wrapper.dart';
@@ -105,7 +105,11 @@ class _MemoriesPhotoSliderViewState extends State<MemoriesPhotoSliderView> {
       return;
     }
 
-    orgMediaService.storedPath.copySync(newMediaService.originalPath.path);
+    if (orgMediaService.storedPath.existsSync()) {
+      orgMediaService.storedPath.copySync(newMediaService.originalPath.path);
+    } else if (orgMediaService.tempPath.existsSync()) {
+      orgMediaService.tempPath.copySync(newMediaService.originalPath.path);
+    }
 
     if (!mounted) return;
